// src/rides/domain/constants/ride-payment-detail-status.enum.ts
export enum ERidePaymentDetailStatus {
  PENDING = 'pending',
  QUEUED = 'queued',
  INITIATED = 'initiated',
  FAILED = 'failed',
  CAPTURE = 'capture',
  SETTLEMENT = 'settlement',
  SUCCESS = 'success',
  COMPLETED = 'completed',
  DENY = 'deny',
  CANCEL = 'cancel',
  CANCELLED = 'cancelled',
  EXPIRED = 'expired',
  EXPIRE = 'expire',
  FAILURE = 'failure',
  AUTHORIZED = 'authorized',
  AUTHORIZE = 'authorize',
  CHALLENGE = 'challenge',
  UNKNOWN = 'unknown',
}